feat: 컨트롤러 인터페이스 구현
컨트롤러 인터페이스 추가
